color: drop "transfer function" from TF descriptions
Several places that print this string already include "transfer
function", so there is no need to say it twice. The one place where it
is needed is fixed accordingly. Other places already imply it enough.

This shortens the prints without losing anything.

A few descriptions are improved to help understanding.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>