backend-drm: define struct drm_kms_device
This gives the anonymous struct at struct drm_device::drm a name.

drm_device::drm is renamed to kms_device, which I think is a better
name, lacking an even better one.

The type of kms_device field is changed into a pointer, so that struct
drm_kms_device could be created and destroyed separately from struct
drm_device. Actually implementing this is left for the next patch, and
here the pointer is temporarily initialized with

	device->kms_device = &device->kms_device_allocd;

Changing the name and type causes tons of trivial changes all over the
DRM-backend. This patch does only that, and the next patch will be more
interesting.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>