gl-renderer: Fix memory leak for mixed opaque/blended views
transform_damage() returns an allocated set of quads; if a surface had
both opaque and blended regions, we were overwriting the
previously-allocated set of quads for the blended region.

Luckily, transform_damage() doesn't need to be called twice anyway, so
we can fix this by only calling it once in the first case.

Signed-off-by: Daniel Stone <daniels@collabora.com>